Document Additional Methods
Provide documentation for additional methods(in particular those not available on MDN for reference) to better explain their purpose and usage. Whilst some of these methods may be private and/or non-exported - this information may still prove useful for library contributors and maintainers(some IDEs can provide this information in a popup dialog box).

Also, make more usage of `{@link }` to link to other classes, functions, variables etc. when talking about them in the docblocks

Ref #32