Encode PNG To Stream
Write test for encoding a PNG to a stream. Though, using a passthrough stream rather than a write stream so that we don't actually have to write anything to disk and we can make assertions about the kind of data that was actually written. Also bringing in the `fileType` module to analyse the magic number in the file header to see what kind of file it is in a custom jest matcher. That way, we can assert that something is a JPG, or that it's a PNG etc.

Ref #43